# docs/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Spektors API",
...
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/chat/completions": {
      "post": {
        "tags": [
          "llm"
        ],
        "summary": "Прокси OpenAI chat/completions → Ollama",
        "description": "Пробрасывает JSON в `{llm_base}/v1/chat/completions`. Требуется заголовок **`X-App-Id`** (как у `/api/v1/chat`). Модель принудительно берётся из конфига приложения (`ollama_model` / дефолт окружения), поле `model` в теле игнорируется — защита от произвольного выбора модели.\n\nПри `stream: true` возвращает `text/event-stream` (как у OpenAI). Тело запроса — в формате OpenAI Chat Completions.",
        "operationId": "chat_completions_api_v1_chat_completions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "messages": {
                    "items": {
                      "properties": {
                        "role": {
                          "type": "string"
                        },
                        "content": {
                          "type": "string"
                        }
                      },
                      "type": "object"
                    },
                    "type": "array",
                    "description": "История в формате OpenAI"
                  },
                  "stream": {
                    "type": "boolean",
                    "description": "true — ответ `text/event-stream` (SSE)",
                    "default": false
                  },
                  "model": {
                    "type": "string",
                    "description": "Игнорируется сервером; оставлено для совместимости с OpenAI SDK",
                    "deprecated": true
                  }
                },
                "type": "object",
                "required": [
                  "messages"
                ],
                "description": "Формат OpenAI Chat Completions. Заголовок **X-App-Id** обязателен (Authorize в Swagger). Поле **model** в теле **игнорируется** — подставляется из `ollama_model` приложения в `config/apps.yaml` (или `DEFAULT_OLLAMA_MODEL`)."
              },
              "example": {
                "messages": [
                  {
                    "role": "user",
                    "content": "Привет"
                  }
                ],
                "stream": false
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "JSON при stream=false; SSE при stream=true",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              },
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "Нет JSON, нет messages или нет X-App-Id"
          },
          "403": {
            "description": "Неизвестный X-App-Id"
          },
          "502": {
            "description": "Ollama недоступен или ошибка прокси"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/chat": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Сообщение в чат (Ollama + PostgreSQL)",
        "description": "Сохраняет реплику пользователя, вызывает модель для текущего `app_id` и пишет ответ ассистента в БД.\n\n**Заголовок `X-App-Id`** обязателен (см. Authorize в Swagger).\n\n- `stream: false` — ответ `application/json` с полями `session_id`, `thread_id`, `message`.\n- `stream: true` — `text/event-stream` (SSE); `session_id` и `thread_id` в заголовках ответа `X-Session-Id` и `X-Thread-Id`.\n\nОпциональные `session_id` / `thread_id` в теле — **hex 8–64** (как UUID без дефисов); иначе **422**.",
        "operationId": "post_chat_api_v1_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "JSON (stream=false) или SSE (stream=true)",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "session_id": {
                      "type": "string"
                    },
                    "thread_id": {
                      "type": "string"
                    },
                    "message": {
                      "additionalProperties": true,
                      "type": "object",
                      "description": "Ответ Ollama в формате OpenAI chat/completions"
                    }
                  },
                  "type": "object",
                  "required": [
                    "session_id",
                    "thread_id",
                    "message"
                  ]
                }
              },
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary",
                  "description": "События в формате OpenAI streaming"
                }
              }
            }
          },
          "400": {
            "description": "Нет тела, неверный JSON или нет X-App-Id"
          },
          "403": {
            "description": "Неизвестный app_id или несовпадение session/thread"
          },
          "422": {
            "description": "Ошибка валидации (в т.ч. формат session_id/thread_id: hex 8–64)"
          },
          "502": {
            "description": "Ollama недоступен"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
    "/api/v1/rag/query": {
      "post": {
        "tags": [
          "rag"
        ],
        "summary": "Поиск по коллекции RAG",
        "description": "**Заголовок `X-App-Id`** обязателен. Строит эмбеддинг запроса через Ollama (`OLLAMA_EMBED_MODEL`), затем поиск в коллекции `chroma_collection` из конфига приложения. Нужны `rag_enabled: true` и непустая коллекция в `apps.yaml`.",
        "operationId": "rag_query_api_v1_rag_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RagQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RagQueryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Нет X-App-Id"
          },
          "403": {
            "description": "RAG отключён для приложения (`rag_disabled`)"
          },
          "404": {
            "description": "Ошибка коллекции Chroma"
          },
          "502": {
            "description": "Сбой эмбеддинга Ollama или запроса к Chroma"
          },
          "503": {
            "description": "Chroma не настроен или недоступен"
          },
          "422": {
            "description": "Ошибка валидации query/top_k"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatRequest": {
        "properties": {
          "content": {
            "type": "string",
            "maxLength": 32000,
            "minLength": 1,
            "title": "Content"
          },
          "stream": {
            "type": "boolean",
            "title": "Stream",
            "default": true
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id",
            "description": "Опционально; hex 8–64, без дефисов. Продолжить существующую сессию."
          },
          "thread_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thread Id",
            "description": "Опционально; hex 8–64. Продолжить существующий тред в рамках сессии."
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "ChatRequest",
        "example": {
          "content": "Привет",
          "session_id": "a1b2c3d4e5f6789012345678abcdef01",
          "stream": true,
          "thread_id": "fedcba0987654321fedcba0987654321"
        }
      },
...
      "RagHit": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "document": {
            "type": "string",
            "title": "Document"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "distance": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance"
          }
        },
        "type": "object",
        "required": [
          "document"
        ],
        "title": "RagHit"
      },
      "RagQueryRequest": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Query"
          },
          "top_k": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Top K",
            "default": 5
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "RagQueryRequest",
        "example": {
          "query": "Как подключить карту?",
          "top_k": 5
        }
      },
      "RagQueryResponse": {
        "properties": {
          "collection": {
            "type": "string",
            "title": "Collection"
          },
          "hits": {
            "items": {
              "$ref": "#/components/schemas/RagHit"
            },
            "type": "array",
            "title": "Hits"
          }
        },
        "type": "object",
        "required": [
          "collection",
          "hits"
        ],
        "title": "RagQueryResponse"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "description": "Идентификатор приложения из `config/apps.yaml` (например `web_client`, `web_lite`, `web_admin`). Обязателен для чата, RAG-query и `POST /api/v1/chat/completions`. В Swagger: кнопка **Authorize** → API Key.",
        "in": "header",
        "name": "X-App-Id"
      },
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Для **`/internal/admin/*`** — значение **`INTERNAL_ADMIN_TOKEN`**; для **`/internal/dev/*`** — **`INTERNAL_DEV_TOKEN`**, если переменная задана на сервере. Иначе заголовок не требуется (только разработка). При **`INTERNAL_AUTH_STRICT`** без настроенного токена — **503**."
      }
    }
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Текущий хост (как в браузере)"
    }
  ]
}